requirements.txt
openai
PyGithub
python-dotenv
pytest
black
fastapi
uvicorn
httpx